api-django/emporium_api_django/requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
